restructure directory of service components.
