+++ signals-service/Dockerfile
@@ -19,7 +19,7 @@ FROM ${BASE} AS builder
 
 ARG MAKE=make build
 
-WORKDIR /device-sdk-go
+WORKDIR /service
 
 LABEL license='SPDX-License-Identifier: Apache-2.0' \
   copyright='Copyright (c) 2023: Intel'
@@ -30,21 +30,22 @@ COPY go.mod vendor* ./
 RUN [ ! -d "vendor" ] && go mod download all || echo "skipping..."
 
 COPY . .
+
+#refer to the Makefile for what gets run here
 RUN ${MAKE}
 
 # Next image - Copy built Go binary into new workspace
 FROM alpine:3.18
-LABEL license='SPDX-License-Identifier: Apache-2.0' \
-  copyright='Copyright (c) 2022: Intel'
+LABEL copyright='Copyright (c) 2023; Outside Labs'
 
 RUN apk add --update --no-cache
 
 WORKDIR /
-COPY --from=builder /device-sdk-go/example/cmd/device-simple/Attribution.txt /Attribution.txt
-COPY --from=builder /device-sdk-go/example/cmd/device-simple/device-simple /device-simple
-COPY --from=builder /device-sdk-go/example/cmd/device-simple/res/ /res
+COPY --from=builder /service/Attribution.txt /Attribution.txt
+COPY --from=builder /service/core/res/ /res
+COPY --from=builder /service/core/signals-service /signals-service
 
 EXPOSE 59999
 
-ENTRYPOINT ["/device-simple"]
+ENTRYPOINT ["/siginals-service"]
 CMD ["-cp=consul.http://edgex-core-consul:8500", "--registry"]
