--- signals-service/cmd/device-simple/res/configuration.yaml
@@ -1,31 +1,0 @@
-Writable:
-  LogLevel: INFO
-  Reading:
-    ReadingUnits: true
-  Telemetry:
-    Metrics: 
-      # All service's custom metric names must be present in this list. All common metric names are in the Common Config
-      ReadCommandsExecuted: true
-Service:
-  Host: "localhost"
-  Port: 59999 # Device service are assigned the 599xx range
-  StartupMsg: device simple started
-
-MessageBus:
-  Optional:
-    ClientId: device-simple
-
-# Example overriding of Common Config settings
-Device:
-  AsyncBufferSize: 1
-  # These have common values (currently), but must be here for service local env overrides to apply when customized
-  ProfilesDir: ./res/profiles
-  DevicesDir: ./res/devices
-  # Only needed if device service implements auto provisioning
-  ProvisionWatchersDir: ./res/provisionwatchers
-# Example structured custom configuration
-SimpleCustom:
-  OnImageLocation: ./res/on.png
-  OffImageLocation: ./res/off.jpg
-  Writable:
-    DiscoverSleepDurationSecs: 10
