--- signals-service/cmd/device-simple/res/devices/simple-device.json.example
@@ -1,18 +1,0 @@
-[
-  {
-    "name": "Simple-Device01",
-    "profileName": "Simple-Device",
-    "description": "Example of Simple Device",
-    "labels": ["industrial"],
-    "protocols": {
-      "other": {
-        "address": "simple01",
-        "port": 300
-      }
-    },
-    "autoEvents": [
-      {"interval": "10s", "onChange": false, "sourceName": "Switch"},
-      {"interval": "30s", "onChange": false, "sourceName": "Image"}
-    ]
-  }
-]
