+++ signals-service/core/res/devices/simple-device.json.example
@@ -1,0 +1,18 @@
+[
+  {
+    "name": "Simple-Device01",
+    "profileName": "Simple-Device",
+    "description": "Example of Simple Device",
+    "labels": ["industrial"],
+    "protocols": {
+      "other": {
+        "address": "simple01",
+        "port": 300
+      }
+    },
+    "autoEvents": [
+      {"interval": "10s", "onChange": false, "sourceName": "Switch"},
+      {"interval": "30s", "onChange": false, "sourceName": "Image"}
+    ]
+  }
+]
